Add value-assumption optimization
That's probably not the name some professor entirely full of himself gave to this type of optimization sometime in the late 60's, however neither do I care, nor do I know who they were, as I didn't read about this anywhere, but came up with it by myself.
Basically, ast-optimization will try to keep track of the values of variables after being assigned constant values, remembering and merging those assumptions at conditional branches. (it gets kind-of complicated when dealing with loops, and becomes entirely impossible when encountering labels)
Anyways, the plan was to enable this optimization at O3, however it causes quite a bit of lag, thus causing me to change my mind and only have it enabled at O4, as well as be disabled when compiling modules