Add a user-code interface for TPP
The plan now is to all the different compiler components exported into user-code, thus allowing for usercode to control every aspect of the compiler (though in an implementation-defined manner). - Using this, I will eventually add a user-code moduler that can be used for compiling and execution deemon code at runtime from strings. (Though it's goting to be a long way)