Continue work on new class system
The runtime aspect is pretty much there. The only thing missing now is DEC, as well as compiler support.
To recap: the new class system merges all of the (previously seperate) class traits (such as name, doc and member tables) into a single class-descriptor structure (think of the descriptor being something like a prototype for instantiating a class). Because of this, class creation is much simpler, and requires way less overhead, as well as allowing for general improvements in operator performance by now allowing user-defined operators to be cached, as well as be defined as part of a class's member table, alongside other member functions.
Note though, that in its current state the new class system can't actually be used yet (which is why it's still disabled, with the old system still in place)