Stricter checks for symbol side-effects
Accessing a property/getset symbol without using its value may look like it doesn't have any side-effects. However because it is connected to a user-defined callback, the access cannot be gotten rid of, because the user expects the callback to still be invoked.
For this purpose, a new test was added /util/test/unused_getset.dee that ensures that the compiler doesn't remove symbol accesses that have side-effects