Rework MRO to fix caching inherited attributes
Additionally to fixing MRO in various special cases, this commit also introduces caching of attributes from user-defined classes.
The changes that were required for this are quite extensive, but the new system now even allows for attributes to be sub-divided by what type implements them as part of the inheritance tree.