Lazily allocate doc strings when querying attributes
Since the doc-strings of attributes implemented in C are written as C-strings, rather than string objects, accessing them requires the allocation of a new string object. This commit changes the behavior such that doc strings aren't allocated when the attribute is accessed, but when the doc is accessed itself (s.a. `/deemon/attribute/doc')