Add a way to untrack heap blocks when they shouldn't appear as leaks
With all the statically defined strings, using any one of them in the context of a custom encoding entails the problem that the lazily allocated view will remain as a would-be memory leak, when in fact it should be viewed as part of the static data segment, only to-be deallocated when the application shuts down, rather than be dumped as a leak.
One big culprit is the `DEEMON_MAXBACKLOG' string from `net', which gets its utf16 view allocated when it is used to lookup an environment variable. However, this view is cached and never deallocated (which is intended!).

This commit adds a new function `DeeDbg_UntrackAlloc()', which is meant to delete the internal tracking point for a given point, causing it to not appear in memory leak dumps.