Introducing Type Annotations!
A huge new sub-system has been added to add both compile-time, as well as run-time support for annotating the types of variables, properties, function argument & return types, as well as overall change how types are encoded in documentation strings.
These changes shouldn't affect the ability of existing code to function properly, but only add the ability to allow a programmer to make the typing of certain variable clear, propagating that information throughout bot the compiler, and all the way into the documentation server (which, alongside the `doc' module have been rewritten from scract).
While on a surface-level, type annotations (or in many places also referred to as declaration information) is mainly intended for debug information (the compiler will _not_ generate any code to assert the typing of variables), the long-term idea is to all the compiler to see type annotations as hints during optimization and code generation, allowing it to assume that the annotation will be met at runtime, with the only requirement to the generated assembly being that an expectation miss-match must not result is a hard-crash, but in weak undefined behavior (i.e. intended behavior / exceptions being thrown)
And although currently type annotations don't change actual generated assembly, this may change in the future.

Annotation information is encoded in documentation strings, meaning that accessing this information at runtime is only possible for symbols that can have doc strings associated with them (i.e. globals and class members/operators, though not local variables of local/anonymous or lambda-like functions)

Access to type-information is made through `Doc from doc', which has been rewritten, with the module now containing numerous helper classes to decode type information and allow a user to access this information. (for an example of this, see the new test-case /util/test/type_annotations.dee)