Allow bound-expressions for items (`foo[10] is bound')
To support this, a new instruction, new runtime functions, and a new symbol exported from `deemon' has been added in order to allow this feature to function.
The basic idea is to handle both KeyError and UnboundItem exceptions during item lookup by simply returning false, thus allowing user-code to test if a given index or key has a valid assigned to it within a sequence- or mapping-like base object.