Start implementing an alternative JIT compiler
Due to the enormous overhead (as well as its inability to operate concurrently with itself because it uses global variables), work has started on a secondary, interpretative compiler that can be used to evaluate deemon user-code in the form of strings through use of a new function `exec from deemon'
The idea behind this addition should be quite clear, as this will allow user-code to directly execute strings as turing-complete deemon code in an interpreted & controlled environment.
Currently, only simple expressions are understood, but there shouldn't really be much to stand in the way of extending this for full support of everything that is deemon.