Fix a bottleneck, causing compilation to be up to 70% faster!
This is most notable for larger input files such as the runtime's larger one `doc.dee', which (when needing to be re-compiled) now takes less than 1/3 of the time it took before.
The problem had to do with how TPP used to count line-feeds in order to provide debug information & compiler warnings. - Essentially, each time `TPPFile_LCAt()' was called, it had to scan _every_ character since the start of the last text chunk, all the way to the text pointer being queried. - And as soon as all that memory no longer fit into the CPU's cache, it would slow down to a crawl, having to reload all of that memory hundreds of times.
The fix is quite simple, in that TPP now caches a single line number & associated text pointer, meaning that when requesting line information, only the text between the last checked location and the new location has to be scanned.