Add a way of enumerating keyword names of functions (`__kwds__')
For user-defined functions, we're simply enumerating `code->co_keywords', and for native functions we extract that information from the first line of the documentation string, or return an empty sequence if that doesn't yield any results, of if the function is implemented to not take any keyword arguments.

Also: Add support for `doc' to make use of __kwds__-meta information to construct documentation and prototype information.