Don't use libc's memmem function
Not only is memmem() documented to be broken across a large range of different libc versions, but apparently it's supposed to return `haystack' when `needle_len' is `0', while deemon expects it to return NULL in that case.
Note however that KOS's memmem() implementation does what we need, so still keep using that one if it's available.