Initial support for yield functions in JIT
Currently, `yield' only works when not recursed by anything other than scopes or if-statements. - However, the system currently in place can and will be expanded to also work with for-, foreach-, while-, do...while-, switch-, try- and with-statements.
Right now, if-statements are merely a kind-of proof of concept if you will...