Finish applying KOS coding conventions (first pass)
This was by far the greatest pass of all. And while there are still a lot of places that need fix-up, most of those can be done by use of regex expressions.
This commit's main purpose is to allow for an easy revert in case I mess up one of those regex expressions.