Begin centralizing system dependencies & feature checks
Also: Already add a whole bunch of new feature checks, each one of which makes deemon that much more portable

Note however that this migration isn't done yet, and that there is a good possibility that deemon won't yet be able to be built on your platform of choice. (Until the migration is finished, I'll probably only be testing each step on msvc, and if there's one thing that microsoft's headers are know for, then it's their inability to keep their namespaces clean; e.g. their `<string.h>' header indirectly, and wrongly defines `va_list')
Anyways: The eventual plan will be to add an option to the configure script to test for the presence of all of these options, though still allow deemon to configure itself at compile-time, so-long as the host platform is recognized.