Fix various problems related to stale references during module lookup
This fixes a race condition between a thread calling gc.collect() on a module not referenced anywhere, while another thread is trying to dynamically import() that same module at a point when the first thread has already decremented its reference counter to 0, but has yet to remove the module from global caches.
In this szenario, the second thread must therefor always use `Dee_IncrefIfNotZero()' (or `tryincref()' as it would be called on KOS), the module, and handle failure in doing so identically to not having found the module in-cache at all.