Design (& start implement) a new syntax and encoding/decoding scheme for documentation texts
This user-visible syntax is mainly based on markdown, and is designed to make writing documentation as pleasent as possible.
The deemon compiler will then try to parse your documentation texts in order to detect special constructs (based on markdown), and then encodes these constructs in a format easier to understand by runtime documentation processors (i.e. the `doc-browser', as well as the backing `doc' module).
Currently, the compilation portion is partially implemented, though implementation of a runtime decoder has yet to begin (though this runtime decoder will hook into the currently unimplemented `doc.TextSegment' class, which has always been meant to take on this very purpose)