Extend the table detection algorithm to allow row-match-by-position
This alternative method is used when a row-line is found with too many, vertical characters ('|'-characters), in which case the parser will try to compare the posistions of |-characters to the corner positions of the most recent table row that it was able to parse correctly. If all of the positions line up correctly, then it will re-use column offets from the previous row in order to forcably re-align table contents.
With this method, table bodies pretty much never have to go so far as to escape |-characters in flow-text (which would have been really annoying if it wasn't for this "fix")