Add a new function `int.bitcount()'
This function returns the number of bits needed to represent the associated integer as base-2 as either a signed-, or unsigned integer.
Additionally, this function is now used to fill in the `length' argument of `int.tobytes()', making the argument optional for that function.
Finally, add a test to verify that bitcount() works in all of the corner cases.