New language feature: java-style lambda syntax
These are syntactic sugar for the already-existing lambda syntax. However, writing lambda functions with this new syntax is more consize when compared to using the old (c++-style) syntax.
But note that this doesn't mean that the old syntax is deprecated per sé. (Though from now on I'll probably be exclusively using the new syntax, and the old syntax may become deprecated eventually).

Also note that there might be some new corner cases where parsing code is now ambiguous, or is done so incorrectly (though I've tried to cover all of the cases I could think of)

Finally, the new syntax has yet to be added to the jit parser; only the main compiler is able to parse it (for now)