Use configure-time test to check if va_list is an array
In a perfect world, hybrid's `__VA_LIST_IS_ARRAY' macro would already tell us this on _all_ platforms. - But that's not the world we live in, and as the author of hybrid, I have to admit that there might be platforms where va_list is an array that I simply havn't encountered yet, havn't realized them as using arrays here, or have forgotten this to be the case for them.
Anyways: another step towards making deemon just that tiny bit more portable!