Add some select math functions & constants to `float'
With this, usercode gets access to `float.inf', and `float.nan', `float.isinf', `float.isnan', and a whole of buch of other functions.

Note that many of these were already available via the `math' dex, but this commit makes them available via object attributes. - Additionally, `int' also implements many of the same attributes, making it easy to write code that works with both float and int.