Functions/class symbols declared like `function foo' are final by default
iow: you can't re-assign the function/class name symbol. This in turn will allow the compiler/optimizer to assume that functions/classes can't be reassigned at runtime, thus allowing it to better predict expression types and the like!

Note that this technically is a breaking change for code like:
> function foo() { print "v1"; }
> function foo() { print "v2"; }
> foo();
Since up until now, you were allowed to simply re-assign variables like that. However, with this commit, such code must be re-written as:
> varying function foo() { print "v1"; }
> varying function foo() { print "v2"; }
> foo();
or:
> local foo = () -> { print "v1"; }
> local foo = () -> { print "v2"; }
> foo();