Implement `environ` in the `posix` dex
Note that with this, there are actually 2 implementations of `environ` in deemon (one in `fs` and one in `posix`). However, `fs` as a dex is scheduled for removal (and is to-be replaced by a regular `*.dee` module that is implemented using `posix`). As such, eventually only the variant from `posix` will remain (and will also be available as `fs.environ`)