Don't print compiler errors from the main module twice
This has been a long-running bug that's always been a thorn in my eyes. -- With this, compiler errors caused while the main module (and its dependencies) are being compiled are no longer dumped when deemon exits. This is because these compiler errors had already been logged as they happend (because when compiling the main module and its dependencies, the compiler is configured to directly log errors as they happen).

This differs from the behavior of compiler errors that happened when compiling a dynamically imported module (i.e. `import(".broken-module")`). Compiler errors from such a module will not be logged immediatly (giving code a chance to `catch` the error), but will still be logged as usual if they end up being unhandled exceptions.