Implement auto-format of `symlink()` text argument
This fixes slashes in the windows impl, since the windows kernel *does* properly save `/` in symlink texts, however is unable to interpret that character as a path separator. Instead, only `\` can be used as a separator, and when calling `posix.symlink()`, forward-slashes are replaced with backward-slashes.

However, this behavior can be circumvented by using `posix._symlink()` instead (similar to how `posix._open()` is a raw `open(2)` system call, while `posix.open()` is an `open(2)` system call with some extra sanitization of input arguments)