New Mapping type `collections.RBTree`
This new type of mapping behaves a little different from other mappings, in that it operates on key-ranges, rather than singular keys. This makes using it a little more complicated than it just being a drop-in replacement for other mapping types.

Not fully implemented yet (it's only possible to insert new elements via `setnew()`, since I've yet to implement the node-split functionality)