More work implementing new sequence operators
With this, you'd actually be able to turn on the new system, and tests would actually succeed. However, I still want to improve integration (especially when it comes to default sequence functions, which pretty much all need to be re-written from scratch), and figure out what to do with NSI operators that can't be emulated using the new sequence operators.