Enable `CONFIG_EXPERIMENTAL_NEW_SEQUENCE_OPERATORS`
The time has come. I've tested everything as much as I can, and everything seems to work with the new system. And this is the next step: enabling it by default.

Backwards compatibility will remain for some more time (in case this causes something to break horribly), but even when that's the case, reverting this will only be a temporary measure if deemon needs to function quickly and I can't fix the problem fast enough.