Re-write `string.format` for more features/performance
Since this function is the core driver of template strings, it stands to reason that it should be as efficient as possible, and be capable of using as many fast-pass runtime features as can be.

Additionally, this adds a whole expression stack to format string arguments, pretty much enabling the use of arbitrarily complex expressions within template string arguments.