Add files for building deemon with `v143`
For some reason, I still have to define an SDK version `10.0.17763.0` on my old setup (which is why I'm keeping it in the old v141 and v142 project files), and the github build action was always able to deal with it also. However, the new `v143` one seems to be able to compile without having to manually specify the toolset version, which is nice.